+++ custom_model_runner/datarobot_drum/drum/main.py
@@ -116,9 +116,12 @@ def main():
                 return
 
             import pydevd_pycharm
-            pydevd_pycharm.settrace('localhost', port=40099, stdoutToServer=True, stderrToServer=True)
 
-            skip_predict = os.environ.get('SKIP_PREDICT')
+            pydevd_pycharm.settrace(
+                "localhost", port=40099, stdoutToServer=True, stderrToServer=True
+            )
+
+            skip_predict = os.environ.get("SKIP_PREDICT")
             if skip_predict is not None:
                 sys.argv.extend([ArgumentsOptions.SKIP_PREDICT])