+++ tests/conftest.py
@@ -67,6 +67,8 @@ from tests.drum.constants import (
     TRANSFORM,
     UNSTRUCTURED,
     XGB,
+    TARGET_NAME_DUPLICATED_X,
+    TARGET_NAME_DUPLICATED_Y,
 )
 
 _datasets = {
@@ -94,6 +96,12 @@ _datasets = {
         TESTS_DATA_PATH, "10k_diabetes_sample.csv"
     ),
     (None, BINARY_SPACES): os.path.join(TESTS_DATA_PATH, "iris_with_spaces.csv"),
+    (SKLEARN_REGRESSION, TARGET_NAME_DUPLICATED_X): os.path.join(
+        TESTS_DATA_PATH, "target_name_duplicated_X.csv"
+    ),
+    (SKLEARN_REGRESSION, TARGET_NAME_DUPLICATED_Y): os.path.join(
+        TESTS_DATA_PATH, "target_name_duplicated_y.csv"
+    ),
 }
 
 _training_models_paths = {