+++ custom_model_runner/datarobot_drum/drum/description.py
@@ -1,3 +1,3 @@
-version = "1.4.13"
+version = "1.4.14"
 __version__ = version
 project_name = "datarobot-drum"