+++ custom_model_runner/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ##### Changes
 - bugfix to prevent resampling of data
 - surface warning but don't error out if prediction consistency check fails
+- SKIP_PREDICT env var in fit.sh to bypass pred checks
 
 #### [1.4.11]
 ##### Changes