+++ custom_model_runner/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - surface warning but don't error out if prediction consistency check fails
 ##### Fixed
 - don't fail on spaces in binary class labels in prediction checks
+- fix error where X has colname '0' and target is unnamed
 
 #### [1.4.11]
 ##### Changes