+++ custom_model_runner/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [1.4.14] - in progress
+##### Changed
+- default perf tests timeout from 180 to 600 s
+
+
 #### [1.4.13] - 2021-02-01
 #### Added
 - custom Java predictors support